Implemented Constant Boolean and Check For it
Only is recursive if the boolean has not yet been set. Note, there may be issues if a change is made to a term/factor/argument that makes it constant, since the outer expression does not auto update. In this case, it is better to call a reset of the isConstant parameter in the outer expression to null so that it recalculates.

For example, if a complex expression simplifies down to a constant after simplification (non constant terms cancel), any outer expression with this inner expression as an argument would have no way of knowing that it is now changed.

Alternatively, we can just ensure that it always recurses to avoid this issue in its entirety